app/requirements.txt
fastapi
uvicorn
langchain-groq
langchain
pydantic
python-dotenv
langchain-core